update to 21w03a, bug fix

+++ src/main/java/org/kilocraft/essentials/extensions/magicalparticles/ParticleAnimationManager.java
@@ -9,6 +9,7 @@ import net.minecraft.particle.*;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.math.Vec3f;
 import net.minecraft.util.registry.Registry;
 import ninja.leaping.configurate.ConfigurationNode;
 import ninja.leaping.configurate.ConfigurationOptions;
@@ -152,7 +153,7 @@ public class ParticleAnimationManager implements ReloadableConfigurableFeature, 
                     }
 
                     if (shouldContinue)
-                        particleEffect = new DustParticleEffect(new Vec3d(
+                        particleEffect = new DustParticleEffect(new Vec3f(
                                 Float.parseFloat(rgb[0]), Float.parseFloat(rgb[1]), Float.parseFloat(rgb[2])), section.scale
                         );
                 } else {
